requirements.txt
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scipy>=1.7.0
scikit-learn>=0.24.0
python-dotenv>=0.19.0
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.5b2
flake8>=3.9.0
pyyaml>=5.4.0 